fix: Construction on small refs with small k
Limit the number of threads used in minimizer
index construction (first step of sshash) to 8.
This avoids some crashes observed when building
the index on very small reference sets or with
very small k / m.  It's unclear why this value
is set as the maximum upstream, but we'll check
on this and set the safe value for now.