[do_tag] v0.13.0
This release finally moves to a piscem-cpp dependency
that uses the offical sshash main branch! This means
that sshash is no longer vendored or based on a fork.

Changes to the sshash index mean that any piscem indices
built with version < 0.13.0 must be rebuilt.

This release also adds a new output file, <BASE>_ver.json
which records the information of the piscem version with
which the index was built, as well as the submodule
commit hashes for the cuttlefish and piscem-cpp dependencies.

Finally, this release updates piscem itself to the
Rust 2024 edition.